fix bug: column 'z' missing from gene_statistics
Could not pipe the results from run_coex_... workflow into
cluster_by_coex_partnership since the former does not add a z-score column into
the gene_statistics data-frame that is passed into it, but the latter required
that a z-score column (named z) was present in order to run.

I added a function to compute the z-score given two-tailed p-values and
direction of change, and a method to call and add the computed z-scores onto
the gene_statistics entry of a CoxpresDbPartners object

Now, within cluster_by_coex_partnership, if z-scores are not pre-existing in
the gene_statistics data-frame, then they are computed and added to the node
data in the graph that is added to the returned CoxpresDbPartners object.