removed some irrelevant tests
Tests exist that check that an error is thrown when trying
to make a CoxpresDbAccessor from a non-data-frame, non-file
and non-archive.

Here we removed two tests. These tests were supposed to
check that an error is thrown when trying to get_gene_ids
from a CoxpresDbAccessor that was built from a non-df,
non-file, non-archive. But the tests were pointless, since
setting up the test data was causing the tests to fail.