Speedy get coex partners (#28)
* styling

* check if a gene is in the 'source_id's

Can only return the coexpression partners of a gene if
there is a file for that gene in the archive, or some
entries for that gene in the `source_id` column of a
dataframe.

This is checked by `get_all_coex_partners`

Also:

- Added function `get_source_ids`

- Rewrote tests on dataframe-based coxpresDbAccessors
  so that they use (exported function)
  `get_coex_partners` rather than (internal function)
  `get_all_coex_partners`

* removed some irrelevant tests

Tests exist that check that an error is thrown when trying
to make a CoxpresDbAccessor from a non-data-frame, non-file
and non-archive.

Here we removed two tests. These tests were supposed to
check that an error is thrown when trying to get_gene_ids
from a CoxpresDbAccessor that was built from a non-df,
non-file, non-archive. But the tests were pointless, since
setting up the test data was causing the tests to fail.

* change filename to reflect class type

Also:

- change @include calls to point to ...CoxpresDbAccessor.R

* refactor tests to contain a single function call

Done to separate setting up test data from running test

* add explicit test against values for 2538791

2538791 is a gene in the test archives that are included as
example data. Previously had tests that checked the
structure of the tables returned by
`get_all_coex_partners`. I copied out the actual values
that are stored in the files for this gene, to reduce the
number of tests.

* add filtering tests for get_coex_partners

Use can filter by mutual-rank or by number of partners

* move invalid-accessor tests to `get_coex_partners`

`get_coex_partners` calls `get_all_coex_partners` and the
latter is not exported; so if a `CoxpresDbAccessor` is
invalid, it should be caught by a test on
`get_coex_partners`

Note that calling get_coex_partners with > 1 gene_id is
allowed, so the check against get_all_coex_partners for
multiple input genes remains in the tests for the latter.

* check genes are valid in `get_coex_partners`

... rather than in `get_all_coex_partners` which is called
by `get_coex_partners`

* use dplyr::top_n to get rows with the lowest mutual-rank (rather than head)

* use group / ungroup in filter_coex_partners

* `.filter_coex_partners` can work on multiple source_ids

* Don't use get_all_coex_partners for a CoxpresDbDataframeAccessor

* remove internal .import_fn and .filter_fn

* make get_all_coex_partners only relevant to CoxpresDbArchiveAccessor objects

Also:
- no need to arrange output of get_all_coex_partners by
  mutual-rank, since this will subsequently be done by
  the calling function `get_coex_partners`

* implement multiple gene_ids in get_all_coex_partners

* check/update docs

* Increment version number